requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
pymupdf==1.24.10
tenacity==9.0.0
python-dotenv==1.0.1